+++ packages/wxt/src/core/builders/vite/index.ts
@@ -20,6 +20,10 @@ import {
 import { Hookable } from 'hookable';
 import { toArray } from '~/core/utils/arrays';
 import { safeVarName } from '~/core/utils/strings';
+import { importEntrypointFile } from '~/core/utils/building';
+import { ViteNodeServer } from 'vite-node/server';
+import { ViteNodeRunner } from 'vite-node/client';
+import { installSourcemapsSupport } from 'vite-node/source-map';
 
 export async function createViteBuilder(
   wxtConfig: ResolvedConfig,
@@ -57,7 +61,6 @@ export async function createViteBuilder(
     config.plugins.push(
       wxtPlugins.download(wxtConfig),
       wxtPlugins.devHtmlPrerender(wxtConfig, server),
-      wxtPlugins.unimport(wxtConfig),
       wxtPlugins.resolveVirtualModules(wxtConfig),
       wxtPlugins.devServerGlobals(wxtConfig, server),
       wxtPlugins.tsconfigPaths(wxtConfig),
@@ -204,21 +207,65 @@ export async function createViteBuilder(
   return {
     name: 'Vite',
     version: vite.version,
-    async importEntrypoint(url) {
-      const baseConfig = await getBaseConfig();
-      const envConfig: vite.InlineConfig = {
-        plugins: [
-          wxtPlugins.webextensionPolyfillMock(wxtConfig),
-          wxtPlugins.removeEntrypointMainFunction(wxtConfig, url),
-        ],
-      };
-      const config = vite.mergeConfig(baseConfig, envConfig);
-      const server = await vite.createServer(config);
-      await server.listen();
-      const runtime = await vite.createViteRuntime(server, { hmr: false });
-      const module = await runtime.executeUrl(url);
-      await server.close();
-      return module.default;
+    async importEntrypoint(path) {
+      switch (wxtConfig.experimental.entrypointImporter) {
+        default:
+        case 'jiti': {
+          return await importEntrypointFile(path);
+        }
+        case 'vite-runtime': {
+          const baseConfig = await getBaseConfig();
+          const envConfig: vite.InlineConfig = {
+            plugins: [
+              wxtPlugins.webextensionPolyfillMock(wxtConfig),
+              wxtPlugins.removeEntrypointMainFunction(wxtConfig, path),
+            ],
+          };
+          const config = vite.mergeConfig(baseConfig, envConfig);
+          const server = await vite.createServer(config);
+          await server.listen();
+          const runtime = await vite.createViteRuntime(server, { hmr: false });
+          const module = await runtime.executeUrl(path);
+          await server.close();
+          return module.default;
+        }
+        case 'vite-node': {
+          const baseConfig = await getBaseConfig();
+          // Disable dep optimization, as recommended by vite-node's README
+          baseConfig.optimizeDeps ??= {};
+          baseConfig.optimizeDeps.noDiscovery = true;
+          baseConfig.optimizeDeps.include = [];
+          const envConfig: vite.InlineConfig = {
+            plugins: [
+              wxtPlugins.webextensionPolyfillMock(wxtConfig),
+              wxtPlugins.removeEntrypointMainFunction(wxtConfig, path),
+            ],
+          };
+          const config = vite.mergeConfig(baseConfig, envConfig);
+          const server = await vite.createServer(config);
+          await server.pluginContainer.buildStart({});
+          const node = new ViteNodeServer(server);
+          installSourcemapsSupport({
+            getSourceMap: (source) => node.getSourceMap(source),
+          });
+          const runner = new ViteNodeRunner({
+            root: server.config.root,
+            base: server.config.base,
+            // when having the server and runner in a different context,
+            // you will need to handle the communication between them
+            // and pass to this function
+            fetchModule(id) {
+              return node.fetchModule(id);
+            },
+            resolveId(id, importer) {
+              return node.resolveId(id, importer);
+            },
+          });
+          const res = await runner.executeFile(path);
+          await server.close();
+          return res.default;
+        }
+      }
     },
     async build(group) {
       let entryConfig;