+++ packages/wxt/src/core/package-managers/__tests__/pnpm.test.ts
@@ -3,17 +3,16 @@ import path from 'node:path';
 import { pnpm } from '../pnpm';
 import { execaCommand } from 'execa';
 
-const cwd = path.resolve(__dirname, 'fixtures/pnpm-project');
-
 process.env.WXT_PNPM_IGNORE_WORKSPACE = 'true';
 
 describe('PNPM Package Management Utils', () => {
-  beforeAll(async () => {
-    // PNPM needs the modules installed, or 'pnpm ls' will return a blank list.
-    await execaCommand('pnpm i --ignore-workspace', { cwd });
-  });
-
   describe('listDependencies', () => {
+    const cwd = path.resolve(__dirname, 'fixtures/simple-pnpm-project');
+    beforeAll(async () => {
+      // PNPM needs the modules installed, or 'pnpm ls' will return a blank list.
+      await execaCommand('pnpm i --ignore-workspace', { cwd });
+    });
+
     it('should list direct dependencies', async () => {
       const actual = await pnpm.listDependencies({ cwd });
       expect(actual).toEqual([